[CLIENT] [WIP] restructure folder structure
