--- Client/src/gameEngine/inc/GameState.hpp
@@ -1,21 +1,0 @@
-/*
-** EPITECH PROJECT, 2023
-** StateMachine
-** File description:
-** GameState
-*/
-
-#ifndef GAMESTATE_HPP_
-#define GAMESTATE_HPP_
-
-class GameState
-{
-public:
-    GameState();
-    ~GameState();
-
-protected:
-private:
-};
-
-#endif /* !GAMESTATE_HPP_ */
